+++ src/app/(house)/cart/page.tsx
@@ -1,17 +1,17 @@
 "use client";
 
-import { CartListContainer } from "@/features/cart/components/CartList/CartListContainer";
 import { withAuth } from "@/shared/hocs/withAuth";
+import { CartPage } from "@/pages/cart/ui/CartPage";
 import * as s from "@/app/(house)/cart/style.css";
 
-function CartPage() {
+function Page() {
   return (
     <main>
       <section className={s.cartSection}>
-        <CartListContainer />
+        <CartPage />
       </section>
     </main>
   );
 }
 
-export default withAuth(CartPage, "protected");
+export default withAuth(Page, "protected");
